// lib/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type AdminImageConnection = {
  __typename?: 'AdminImageConnection';
  /** A list of edges. */
  edges?: Maybe<Array<AdminImageEdge>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<AdminImage>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Category = {
  __typename?: 'Category';
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  fullPath: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  rankable: Scalars['Boolean']['output'];
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  videos?: Maybe<Array<OriginalVideo>>;
};
// ...
export type Comment = {
  __typename?: 'Comment';
  contents?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  currentUserHasLike?: Maybe<Scalars['Boolean']['output']>;
  discardedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  id: Scalars['ID']['output'];
  isMine?: Maybe<Scalars['Boolean']['output']>;
  isSpoiler: Scalars['Boolean']['output'];
  likeNum?: Maybe<Scalars['Int']['output']>;
  replies?: Maybe<ReplyConnection>;
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  user?: Maybe<User>;
  userId: Scalars['Int']['output'];
  videoId: Scalars['Int']['output'];
};
// ...
export type CommentRepliesConnection = {
  __typename?: 'CommentRepliesConnection';
  /** A list of edges. */
  edges?: Maybe<Array<CommentRepliesEdge>>;
  id: Scalars['ID']['output'];
  /** A list of nodes. */
  nodes?: Maybe<Array<Reply>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Coupon = {
  __typename?: 'Coupon';
  discountMonth: Scalars['Int']['output'];
  discountRate: Scalars['Float']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  paymentMethodSkippable: Scalars['Boolean']['output'];
  promotionCodes: Array<PromotionCode>;
  redeemBy: Scalars['ISO8601Date']['output'];
  redeemFrom?: Maybe<Scalars['ISO8601Date']['output']>;
};

/** The connection type for Coupon. */
export type CouponConnection = {
  __typename?: 'CouponConnection';
  /** A list of edges. */
  edges?: Maybe<Array<CouponEdge>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Coupon>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type HomeScreen = {
  __typename?: 'HomeScreen';
  category?: Maybe<Category>;
  country?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  id: Scalars['ID']['output'];
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  videos?: Maybe<Array<OriginalVideo>>;
};
// ...
export type OriginalVideo = Video & {
  __typename?: 'OriginalVideo';
  categories?: Maybe<Array<Category>>;
  comment?: Maybe<Comment>;
  commentaryUrl?: Maybe<Scalars['String']['output']>;
  comments?: Maybe<CommentConnection>;
  commentsCount: Scalars['Int']['output'];
  contentEnd?: Maybe<Duration>;
  cookies?: Maybe<Cookies>;
  createdAt: Scalars['ISO8601DateTime']['output'];
  creator?: Maybe<Creator>;
  currentUserHasLike?: Maybe<Scalars['Boolean']['output']>;
  currentUserVideoRating?: Maybe<VideoRating>;
  description?: Maybe<Scalars['String']['output']>;
  drmDashFileUrl?: Maybe<Scalars['String']['output']>;
  drmHlsFileUrl?: Maybe<Scalars['String']['output']>;
  duration: Duration;
  fileUrl?: Maybe<Scalars['String']['output']>;
  filmarksUrl?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  jiqooNextVideo?: Maybe<OriginalVideo>;
  landscapeThumbnail?: Maybe<Scalars['String']['output']>;
  likeNum: Scalars['Int']['output'];
  nextVideo?: Maybe<OriginalVideo>;
  playlists?: Maybe<Array<Playlist>>;
  playlistsVideos?: Maybe<Array<PlaylistsVideo>>;
  portraitThumbnail?: Maybe<Scalars['String']['output']>;
  publishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  recommends?: Maybe<Array<OriginalVideo>>;
  satisfiedNum: Scalars['Int']['output'];
  status: VideoStatusEnum;
  subtitles?: Maybe<Array<SrtSubtitle>>;
  thumbnailCategory?: Maybe<Category>;
  title?: Maybe<Scalars['String']['output']>;
  trailer?: Maybe<TrailerVideo>;
  translations: Array<VideoTranslation>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  viewingTime?: Maybe<VideoViewingTime>;
  vtts?: Maybe<Array<VttSubtitle>>;
  willBePublishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  willBeUnpublishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
};


export type OriginalVideoCommentArgs = {
  id: Scalars['ID']['input'];
};


export type OriginalVideoCommentsArgs = {
  after?: InputMaybe<Scalars['String']['input']>;
  before?: InputMaybe<Scalars['String']['input']>;
  first?: InputMaybe<Scalars['Int']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
};

/** The connection type for OriginalVideo. */
export type OriginalVideoConnection = {
  __typename?: 'OriginalVideoConnection';
  /** A list of edges. */
  edges?: Maybe<Array<OriginalVideoEdge>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<OriginalVideo>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};

/** An edge in a connection. */
export type OriginalVideoEdge = {
  __typename?: 'OriginalVideoEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<OriginalVideo>;
};

/** Information about pagination in a connection. */
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  adminCoupons: CouponConnection;
  adminImages: AdminImageConnection;
  adminOriginalVideos: OriginalVideoConnection;
  adminScreenPopup?: Maybe<ScreenPopup>;
  adminScreenPopups: ScreenPopupConnection;
  adminTestMovies: Array<TestMovie>;
  adminUsers: UserConnection;
  allOriginalVideos: Array<OriginalVideo>;
  category: Category;
  commentReplies: CommentRepliesConnection;
  coupon?: Maybe<Coupon>;
  currentUser?: Maybe<User>;
  firstRecommendationRow: RecommendationRow;
  homeScreens: Array<HomeScreen>;
  jiqooCategory: Category;
  jiqooCommentReplies: CommentRepliesConnection;
  jiqooHomeScreens: Array<HomeScreen>;
  jiqooOriginalVideo?: Maybe<OriginalVideo>;
  jiqooRootCategories: Array<RootCategory>;
  jiqooScreenPopups: ScreenPopupConnection;
  jiqooTrailerVideo?: Maybe<TrailerVideo>;
  jiqooTrailerVideos: TrailerVideoConnection;
  jiqooVideoComments: VideoCommentsConnection;
  jiqooVideoRecommends?: Maybe<VideoRecommendsConnection>;
  notifications: UserNotificationConnection;
  opsToggles: OpsTogglesResult;
  originalVideo?: Maybe<OriginalVideo>;
  payment: Payment;
  playlist: Playlist;
  playlists: Array<Playlist>;
  promotionCode?: Maybe<PromotionCode>;
  recommendedSearchKeywords: Array<RecommendedSearchKeyword>;
  rootCategories: Array<RootCategory>;
  screenPopups: ScreenPopupConnection;
  trailerVideo?: Maybe<TrailerVideo>;
  trailerVideos: TrailerVideoConnection;
  trialVideos: Array<OriginalVideo>;
  tvSearchCategories: Array<TvSearchCategory>;
  userNotificationSettings: Array<UserNotificationSetting>;
  videoAiRecommends?: Maybe<VideoRecommendsConnection>;
  videoComments: VideoCommentsConnection;
  videoHistories: VideoHistoryConnection;
  videoRatings: VideoRatingConnection;
  videoRecommends?: Maybe<VideoRecommendsConnection>;
  withdrawalVideos: Array<TrailerVideo>;
};
// ...
export type RecommendationRow = {
  __typename?: 'RecommendationRow';
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  videos?: Maybe<Array<OriginalVideo>>;
};
// ...
export type RootCategory = {
  __typename?: 'RootCategory';
  categories?: Maybe<Array<Category>>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  fullPath: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
};

export type ScreenPopup = {
  __typename?: 'ScreenPopup';
  button?: Maybe<ScreenPopupButton>;
  country: Scalars['String']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  displayMethod: ScreenPopupDisplayMethod;
  id: Scalars['ID']['output'];
  image: Scalars['String']['output'];
  isSensitive: Scalars['Boolean']['output'];
  mobileClickCount: Scalars['Int']['output'];
  mobileShowCount: Scalars['Int']['output'];
  pcImage?: Maybe<Scalars['String']['output']>;
  publicationFrom: Scalars['ISO8601DateTime']['output'];
  publicationStatus: Scalars['String']['output'];
  publicationTo: Scalars['ISO8601DateTime']['output'];
  rowOrderRank: Scalars['Int']['output'];
  title: Scalars['String']['output'];
  translations: Array<ScreenPopupTranslation>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  webClickCount: Scalars['Int']['output'];
  webShowCount: Scalars['Int']['output'];
};
// ...
export type ScreenPopupConnection = {
  __typename?: 'ScreenPopupConnection';
  /** A list of edges. */
  edges?: Maybe<Array<ScreenPopupEdge>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<ScreenPopup>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type TestMovie = {
  __typename?: 'TestMovie';
  id: Scalars['ID']['output'];
  locale: Scalars['String']['output'];
  testDescription?: Maybe<Scalars['String']['output']>;
  title: Scalars['String']['output'];
};
// ...
export type User = {
  __typename?: 'User';
  age?: Maybe<Scalars['Int']['output']>;
  autoplayVideos: Scalars['Boolean']['output'];
  avatar?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  dateOfBirth?: Maybe<Scalars['ISO8601Date']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  freeTrialEndDate?: Maybe<Scalars['ISO8601Date']['output']>;
  gender?: Maybe<Scalars['String']['output']>;
  hasCancelledSubscription?: Maybe<Scalars['Boolean']['output']>;
  hasSubscription?: Maybe<Scalars['Boolean']['output']>;
  histories?: Maybe<OriginalVideoConnection>;
  id: Scalars['ID']['output'];
  isSelf?: Maybe<Scalars['Boolean']['output']>;
  likes?: Maybe<LikeConnection>;
  mailDesired?: Maybe<Scalars['Boolean']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  payment: Payment;
  points?: Maybe<Scalars['Int']['output']>;
  qualitySetting?: Maybe<Scalars['Int']['output']>;
  speedSetting?: Maybe<Scalars['String']['output']>;
  stripeCustomerId?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  userStatus?: Maybe<Scalars['String']['output']>;
  withdrawal?: Maybe<Withdrawal>;
};
// ...
export type UserConnection = {
  __typename?: 'UserConnection';
  /** A list of edges. */
  edges?: Maybe<Array<UserEdge>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<User>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type VideoCommentsConnection = {
  __typename?: 'VideoCommentsConnection';
  allCount?: Maybe<Scalars['Int']['output']>;
  /** A list of edges. */
  edges?: Maybe<Array<VideoCommentsEdge>>;
  id: Scalars['ID']['output'];
  /** A list of nodes. */
  nodes?: Maybe<Array<Comment>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};

/** An edge in a connection. */
export type VideoCommentsEdge = {
  __typename?: 'VideoCommentsEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Comment>;
};
// ...
export type GetVideoCommentsQuery = { __typename?: 'Query', videoComments: { __typename?: 'VideoCommentsConnection', id: string, allCount?: number | null, pageInfo?: { __typename?: 'PageInfo', startCursor?: string | null, endCursor?: string | null, hasNextPage: boolean, hasPreviousPage: boolean } | null, edges?: Array<{ __typename?: 'VideoCommentsEdge', cursor: string, node?: { __typename?: 'Comment', id: string, contents?: string | null, createdAt?: any | null, likeNum?: number | null, user?: { __typename?: 'User', id: string, name?: string | null, avatar?: string | null } | null } | null }> | null } };